Address code review: explicit float cast for exposure, toInt validation for resolution combo
Agent-Logs-Url: https://github.com/srcejon/sdrangel/sessions/dc91738b-2ea7-40a9-9b63-705b97fafee2

+++ plugins/feature/camera/cameraworker.cpp
@@ -498,7 +498,7 @@ void CameraWorker::setupQtCapture()
     }
 
     m_qtCamera->setExposureMode(QCamera::ExposureManual);
-    m_qtCamera->setManualExposureTime(m_settings.m_exposureTimeMs / 1000.0f);
+    m_qtCamera->setManualExposureTime(static_cast<float>(m_settings.m_exposureTimeMs) / 1000.0f);
     m_qtCamera->setManualIsoSensitivity(m_settings.m_isoSensitivity);
 
     m_captureSession->setCamera(m_qtCamera);
